backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.4.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.4
cron_descriptor==2.0.6
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.9.0
django-filter==25.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.1
Faker==40.8.0
flake8==7.1.1
flower==2.0.1
gunicorn==25.1.0
humanize==4.15.0
iniconfig==2.3.0
kombu==5.6.2
mccabe==0.7.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.11.0
pytest==8.3.4
pytest-cov==6.1.1
pytest-django==4.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
redis==5.0.1
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
vine==5.1.0
wcwidth==0.6.0
